# services/api/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Shipped API
  version: "1.0.0-phase2"
  description: >
    The Shipped control plane. Most /v1 routes require a Bearer EdDSA JWT
    (issued by Better Auth) which the Go API verifies itself (the authz boundary).
    Every request runs under a Postgres RLS tenant context derived from the
    verified org_id/user_id claims.


    Phase 2 adds access control & domains: the /authz viewer exchange (mint +
    password) which issues a host-scoped EdDSA edge token the serving Worker
    verifies (public JWKS at /.well-known/edge-jwks), per-site access-mode + policy
    management (admin/owner only), allowlist CRUD, the org allow_external_sharing
    toggle (with reconcile), member listing, and Cloudflare-for-SaaS custom domains.
  license:
    name: FSL-1.1-Apache-2.0
servers:
  - url: https://api.shipped.app
    description: Production
  - url: http://localhost:8080
    description: Local / self-host
security:
  - bearerAuth: []
paths:
...
  /v1/sites:
    post:
      summary: Create a site
      description: >
        Reserved slugs are rejected (400). A slug whose content host
        (`<slug>.shippedusercontent.com`) is already taken by another org/site
        returns 409 Conflict — the host registry is global even though site slugs
        are unique only per org. On the cloud build, exceeding the per-user site
        cap returns 402 with the quota body so the dashboard can open the upgrade
        modal.
      operationId: createSite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:
                  type: string
                  description: Subdomain label under shippedusercontent.com (unique per org).
                  example: my-docs
      responses:
        "201":
          description: Site created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/QuotaExceeded" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      summary: List the caller org's sites
      operationId: listSites
      responses:
        "200":
          description: The org's sites
          content:
            application/json:
              schema:
                type: object
                properties:
                  sites:
                    type: array
                    items: { $ref: "#/components/schemas/Site" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/sites/{id}:
    get:
      summary: Get one site
      operationId: getSite
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      responses:
        "200":
          description: The site
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Site" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/sites/{id}/publish:
    post:
      summary: Publish (or roll back to) a version
      description: >
        Flips the site's live-version pointer to version_id and writes the edge
        routing projection (KV route:<host>). Rollback is just publishing an older
        version_id. The pointer flip is Postgres-authoritative; the projection is a
        rebuildable cache.
      operationId: publish
      parameters:
        - { $ref: "#/components/parameters/SiteID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version_id]
              properties:
                version_id: { type: string, format: uuid }
      responses:
        "200":
          description: Published; the site is live
          content:
            application/json:
              schema:
                type: object
                properties:
                  live_url: { type: string, format: uri }
                  version_id: { type: string, format: uuid }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived EdDSA JWT (Better Auth) or a deploy token.
  parameters:
    SiteID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: app.sites.id
  schemas:
...
    Site:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        slug: { type: string }
        owner_id: { type: string, format: uuid }
        access_mode: { type: string, enum: [public, password, allowlist, org_only] }
        current_version_id: { type: string, format: uuid, nullable: true }
        live_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error: { type: string, description: Stable machine-readable code }
        message: { type: string }
    QuotaExceeded:
      type: object
      description: 402 body (see internal/quota.ExceededError); the dashboard opens the upgrade/sales modal.
      properties:
        limit: { type: string, enum: [sites_per_user, members_per_org] }
        current: { type: integer, format: int64 }
        max: { type: integer, format: int64 }
        plan_tier: { type: string }
        next_tier: { type: string }
        upgrade_url: { type: string, format: uri }
        sales_url: { type: string, format: uri }
  responses:
    Unauthorized:
      description: Missing/invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Malformed request / reserved slug / verification failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource absent (or invisible under the caller's tenant)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Slug/host already in use by another org/site (global host registry)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    QuotaExceeded:
      description: A hard cap was crossed (cloud build only)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/QuotaExceeded" }